# docs/swagger.yaml
definitions:
...
  feedback_help.ChangeQuestionReq:
    properties:
      answer:
        type: string
      question:
        type: string
      question_id:
        type: integer
    type: object
  feedback_help.CreateQuestionReq:
    properties:
      answer:
        type: string
      question:
        type: string
    type: object
  feedback_help.DeleteQuestionReq:
    properties:
      question_id:
        type: integer
    type: object
  feedback_help.FindQuestionsByNameResp:
    properties:
      questions:
        items:
          $ref: '#/definitions/feedback_help.FrequentlyAskedQuestion'
        type: array
    type: object
  feedback_help.FrequentlyAskedQuestion:
    properties:
      answer:
        type: string
      clickTimes:
        description: |-
          Utime      time.Time
          Ctime      time.Time
        type: integer
      id:
        type: integer
      question:
        type: string
    type: object
  feedback_help.GetQuestionsResp:
    properties:
      questions:
        items:
          $ref: '#/definitions/feedback_help.FrequentlyAskedQuestion'
        type: array
    type: object
  grade.GetGradeByTermResp:
    properties:
      grades:
        description: 课程信息
        items:
          $ref: '#/definitions/grade.Grade'
        type: array
    type: object
  grade.GetGradeScoreResp:
    properties:
      type_of_grade_scores:
        items:
          $ref: '#/definitions/grade.TypeOfGradeScore'
        type: array
    type: object
  grade.Grade:
    properties:
      Kclbmc:
        description: 课程类别名称
        type: string
      cj:
        description: 最终成绩
        type: number
      finalGrade:
        description: 期末成绩分数
        type: number
      finalGradePercent:
        description: /期末成绩占比
        type: string
      jd:
        description: 绩点
        type: number
      kcbj:
        description: 课程标记(主修/辅修)
        type: string
      kcmc:
        description: 课程名称
        type: string
      kcxzmc:
        description: 课程性质名称
        type: string
      regularGrade:
        description: 平时成绩分数
        type: number
      regularGradePercent:
        description: 平时成绩占比
        type: string
      xf:
        description: 学分
        type: number
    type: object
  grade.GradeScore:
    properties:
      kcmc:
        description: 课程名称
        type: string
      xf:
        description: 学分
        type: number
    type: object
  grade.TypeOfGradeScore:
    properties:
      grade_score_list:
        items:
          $ref: '#/definitions/grade.GradeScore'
        type: array
      kcxzmc:
        description: 课程性质名称
        type: string
    type: object
  infoSum.DelInfoSumRequest:
    properties:
      id:
        type: integer
    type: object
...
  infoSum.SaveInfoSumRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      link:
        type: string
      name:
        type: string
    type: object
...
  user.LoginByCCNUReq:
    properties:
      password:
        description: 密码
        type: string
      student_id:
        type: string
    type: object
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /InfoSum/delInfoSum:
    delete:
      consumes:
      - application/json
      description: 删除信息整合信息
      parameters:
      - description: 删除信息整合信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/infoSum.DelInfoSumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除信息整合信息
      tags:
      - 信息整合
...
  /InfoSum/saveInfoSum:
    post:
      consumes:
      - application/json
      description: 保存信息整合信息,id是可选字段,如果有就是替换原来的列表里的,如果没有就是存储新的值
      parameters:
      - description: 保存信息整合信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/infoSum.SaveInfoSumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 保存信息整合信息
      tags:
      - 信息整合
...
  /feedback_help/changeQuestion:
    post:
      consumes:
      - application/json
      description: 修改一个常见问题的内容与答复
      parameters:
      - description: 修改常见问题
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feedback_help.ChangeQuestionReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      summary: 修改一个问题与答复
      tags:
      - 帮助与反馈
  /feedback_help/createQuestion:
    post:
      consumes:
      - application/json
      description: 创建一个常见问题的内容与答复
      parameters:
      - description: 创建一个常见问题
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feedback_help.CreateQuestionReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      summary: 创建一个问题与答复
      tags:
      - 帮助与反馈
  /feedback_help/deleteQuestion:
    post:
      consumes:
      - application/json
      description: 删除一个常见问题的内容与答复
      parameters:
      - description: 删除常见问题
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feedback_help.DeleteQuestionReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除一个问题与答复
      tags:
      - 帮助与反馈
  /feedback_help/findQuestionsByName:
    post:
      consumes:
      - application/json
      description: 对常见问题进行模糊搜索
      parameters:
      - description: 问题名称
        in: query
        name: question
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                date:
                  $ref: '#/definitions/feedback_help.FindQuestionsByNameResp'
              type: object
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      summary: 搜取问题
      tags:
      - 帮助与反馈
  /feedback_help/getQuestion:
    get:
      consumes:
      - application/json
      description: 获取点击数量最多的10个常见问题
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/feedback_help.GetQuestionsResp'
              type: object
        "500":
          description: 系统异常
          schema:
            $ref: '#/definitions/web.Response'
      summary: 获取常见问题
      tags:
      - 帮助与反馈
  /grade/getGradeByTerm:
    get:
      consumes:
      - application/json
      description: 根据学年号和学期号获取用户的成绩
      parameters:
      - description: 学年号（如 2023 表示 2023~2024 学年）
        in: query
        name: xnm
        required: true
        type: integer
      - description: 学期号（1 表示第一学期，2 表示第二学期）
        in: query
        name: xqm
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回学年和学期的成绩信息
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/grade.GetGradeByTermResp'
              type: object
        "500":
          description: 系统异常，获取失败
          schema:
            $ref: '#/definitions/web.Response'
      summary: 查询按学年和学期的成绩
      tags:
      - 成绩
  /grade/getGradeScore:
    get:
      consumes:
      - application/json
      description: 查询学分
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回学分
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/grade.GetGradeScoreResp'
              type: object
        "500":
          description: 系统异常，获取失败
          schema:
            $ref: '#/definitions/web.Response'
      summary: 查询学分
      tags:
      - 成绩
...
  /users/login_ccnu:
    post:
      consumes:
      - application/json
      description: 通过学号和密码进行登录认证
      parameters:
      - description: 登录请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.LoginByCCNUReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.Response'
      summary: ccnu登录
      tags:
      - 用户
...
  /users/refresh_token:
    get:
      consumes:
      - application/json
      description: 通过长token刷新短token
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.Response'
      security:
      - Bearer Auth: []
      summary: 刷新短token
      tags:
      - 用户
...
swagger: "2.0"